split list dataService functions into own file
