+++ src/services/dataService/__tests__/user.test.js
@@ -2,7 +2,7 @@ const test = require('../../../test')
 const seedData = require('../../../test/seedData')
 const {getUser, getList, findLists} = require('../../dataService')
 
-test('dataService', function(t) {
+test.skip('dataService', function(t) {
   return seedData().then(() => {
     t.test('Get user test', function (st) {
       getUser(1).then(function (user) {
