+++ src/services/dataService/index.js
@@ -1,36 +1,4 @@
-const {List} = require('./db')
+const lists = require('./lists')
 const listItems = require('./listItems')
 
-function findLists() {
-  return List.findAll()
-}
-
-function getList(id) {
-  return List.findOne({
-    where: {id},
-  })
-}
-
-function createList(data) {
-  return List.create(data)
-}
-
-function updateList(id, data) {
-  return List.update(data, {where: {id}, returning: true})
-    .then(updatedLists => {
-      const [numRowsUpdated, updatedRows] = updatedLists
-      return updatedRows[0]
-    })
-}
-
-function deleteList(id) {
-  return List.destroy({where: {id}})
-}
-
-module.exports = {
-  findLists,
-  getList,
-  createList,
-  updateList,
-  deleteList,
-}, listItems)
+module.exports = Object.assign({}, lists, listItems)
